croppulse_backend/requirements.txt
# Core
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9

# Satellite & GIS
earthengine-api==0.1.384
google-cloud-storage==2.13.0
rasterio==1.3.9
shapely==2.0.2
geojson==3.1.0

# Voice & AI
openai-whisper==20231117
openai==1.3.7
langdetect==1.0.9

# Climate
requests==2.31.0
pandas==2.1.3
numpy==1.26.2

# Third-party Integrations
africastalking==1.2.5
twilio==8.11.0
firebase-admin==6.3.0
sendgrid==6.11.0

# Auth & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
cryptography==41.0.7

# Payments
requests-oauthlib==1.3.1  # M-Pesa

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.38.0

# Add to your existing requirements.txt

# Security & Validation
phonenumbers==8.13.26
validate-email==1.3
python-magic==0.4.27

# Rate Limiting
django-ratelimit==4.1.0

# Caching
django-redis==5.4.0

# Monitoring
django-silk==5.0.4  # SQL query profiling

# Geospatial
geopy==2.4.1

# File handling
python-magic==0.4.27